flaskDemo/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.0.2
Flask-SocketIO==3.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pipreqs==0.4.9
python-engineio==3.1.2
python-socketio==3.0.0
requests==2.21.0
six==1.12.0
socketIO-client==0.7.2
urllib3==1.24.1
websocket-client==0.54.0
Werkzeug==0.14.1
yarg==0.1.9
